chore(member): reorganized schedule hook folder structure (#152)
