+++ apps/member/src/components/calendar/CalendarStatusSection/CalendarStatusSection.tsx
@@ -4,7 +4,7 @@ import Section from '@components/common/Section/Section';
 import StatusCard from '@components/common/StatusCard/StatusCard';
 
 import { useSchedule } from '@hooks/queries';
-import { useScheduleCollect } from '@hooks/queries/useScheduleCollect';
+import { useScheduleCollect } from '@hooks/queries';
 import { calculateDDay, findClosestEvent, transformEvents } from '@utils/date';
 
 const CalendarStatusSection = () => {
