+++ apps/member/src/hooks/queries/index.ts
@@ -3,6 +3,7 @@ export * from './board';
 export * from './book';
 export * from './book-loan-record';
 export * from './comment';
+export * from './schedule';
 export * from './queryClient';
 export * from './useAccusesMutation';
 export * from './useBirthday';
@@ -19,7 +20,4 @@ export * from './useMyNotifications';
 export * from './useMyProfile';
 export * from './useNews';
 export * from './usePosts';
-export * from './useSchedule';
-export * from './useScheduleAddMutation';
-export * from './useScheduleDeleteMutation';
 export * from './useUserInfoMutation';
